Backend/test_security.py
#!/usr/bin/env python3
"""
Security Test - Verify API Key Loading and Protection
"""

import os
import sys
from dotenv import load_dotenv

def test_environment_loading():
    """Test that environment variables are loaded correctly"""
    print("🔐 Testing Secure Environment Configuration")
    print("=" * 50)
    
    # Load environment variables
    load_dotenv()
    
    # Test OpenRouter API key
    openrouter_key = os.getenv("OPENROUTER_API_KEY")
    if openrouter_key:
        print("✅ OpenRouter API key loaded from environment")
        print(f"   Key starts with: {openrouter_key[:10]}...")
        print(f"   Key length: {len(openrouter_key)} characters")
        
        # Verify it's the correct format
        if openrouter_key.startswith("sk-or-v1-"):
            print("✅ API key format is correct")
        else:
            print("❌ API key format is incorrect")
    else:
        print("❌ OpenRouter API key not found in environment")
    
    # Test other environment variables
    other_vars = [
        "OPENROUTER_BASE_URL",
        "HOST", 
        "PORT",
        "DEBUG",
        "SECRET_KEY",
        "LOG_LEVEL"
    ]
    
    print("\n🔧 Other Environment Variables:")
    for var in other_vars:
        value = os.getenv(var)
        if value:
            if var == "SECRET_KEY":
                print(f"   ✅ {var}: {value[:10]}... (hidden)")
            else:
                print(f"   ✅ {var}: {value}")
        else:
            print(f"   ⚠️ {var}: Not set")

def test_security_practices():
    """Test security practices"""
    print("\n🛡️ Security Practices Check:")
    print("=" * 30)
    
    # Check if .env file exists
    if os.path.exists(".env"):
        print("✅ .env file exists")
    else:
        print("❌ .env file not found")
    
    # Check if .env.example exists
    if os.path.exists(".env.example"):
        print("✅ .env.example template exists")
    else:
        print("❌ .env.example template not found")
    
    # Check if .gitignore includes .env
    if os.path.exists("../.gitignore"):
        with open("../.gitignore", "r") as f:
            gitignore_content = f.read()
            if ".env" in gitignore_content:
                print("✅ .env files are in .gitignore")
            else:
                print("❌ .env files not in .gitignore")
    else:
        print("⚠️ .gitignore not found")

def test_api_key_validation():
    """Test API key validation in code"""
    print("\n🔍 API Key Validation Test:")
    print("=" * 30)
    
    try:
        # Add app to path
        sys.path.append(os.path.join(os.getcwd(), 'app'))
        
        # Test model manager import
        from app.services.model_manager import OPENROUTER_API_KEY
        
        if OPENROUTER_API_KEY:
            print("✅ Model manager loads API key correctly")
            print(f"   Key validation: {'✅ Valid' if OPENROUTER_API_KEY.startswith('sk-or-v1-') else '❌ Invalid'}")
        else:
            print("❌ Model manager API key is None")
            
    except ImportError as e:
        print(f"⚠️ Could not import model manager: {e}")
    except Exception as e:
        print(f"❌ Error testing API key validation: {e}")

if __name__ == "__main__":
    print("🚀 InfluenceOS Security Test Suite")
    print("=" * 50)
    
    test_environment_loading()
    test_security_practices()
    test_api_key_validation()
    
    print("\n" + "=" * 50)
    print("🎯 Security Test Complete")
    print("\n💡 Security Status:")
    print("   • API keys are loaded from environment variables")
    print("   • Sensitive data is not hardcoded in source code")
    print("   • .env files are excluded from version control")
    print("   • Proper validation and error handling in place")
    print("\n🔐 Your InfluenceOS is securely configured!")
