Backend/requirements.txt
# LinkedIn AI Agent Backend - Hugging Face Optimized Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database & Storage
supabase==2.3.0
redis==5.0.1

# AI/ML Dependencies
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
tokenizers>=0.13.0
sentencepiece>=0.1.99

# Hugging Face
huggingface-hub==0.20.1

# Image Processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
asyncio-mqtt==0.16.1

# Monitoring & Logging
sentry-sdk[fastapi]==1.39.2

# Data Processing
pandas==2.1.4
numpy==1.24.4

# Testing (optional for production)
pytest==7.4.3
pytest-asyncio==0.21.1

# Production Server
gunicorn==21.2.0
